first version of fdmux, for multiple logical connections over one fd.
each write is encapsulated in a packet.  preceded with a length, a
channel, and an indication that it's data.  control messages can
also be sent, to open/close channels, send error messages, set
priorities, etc.

fdmux provides access to the channels over normal file descriptors,
made with file2chan.  reads & writes block until data is available
or until data can be written.  for each channel, a window is
maintained for both directions.  the window is the amount of data
the other side is willing to read.  this ensures the connection
will never block.